bugfix: error in DryRun when creating new folders
It's supposed to remember that the user chosee to create a folder before
so it doesn't ask again for the same folder during the dry run. This
behavior was broken because of an incorrect data type - dict instead of
set. The data type is now corrected, so an exception is not raised when
simulating creating a folder during dry run.